requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
openpyxl==3.1.2
xlrd==2.0.1
reportlab==4.1.0
pandas==2.2.2
python-dotenv==1.0.1
httpx==0.27.0
sqlalchemy==2.0.25
